==> notebooks/QBGPT/0_dataset_creation_2021_data_bowl.ipynb <==
!pip install --upgrade polars
!pip install nfl_data_py
!pip install --upgrade numpy

import os
import pandas as pd
import polars as pl
import nfl_data_py as nfl

from google.colab import drive
drive.mount('/content/gdrive')

os.chdir("/content/gdrive/MyDrive/NFL_Challenge/NFL-GPT/NFL data")

os.listdir()

**2021 Data Bowl**

path = "data bowl 2021"
os.listdir(path)

data = pl.read_parquet(path+"/tracking.gzip")
plays = pl.read_csv(path+"/plays.csv")
players = pl.read_csv(path+"/players.csv")
games = pl.read_csv(path+"/games.csv")

game_adapted = (games.
                select(["gameId", "homeTeamAbbr", "visitorTeamAbbr"]).
                melt("gameId", ["homeTeamAbbr", "visitorTeamAbbr"], variable_name = "team", value_name = "TeamAbbr").
                with_columns(pl.col('team').str.replace("TeamAbbr", '')))

play_adapted = (plays.
                select(["gameId", "playId", "possessionTeam"]).
                rename({"possessionTeam": "TeamAbbr"}).
                with_columns(pl.lit("Offense").alias("OffDef")))

game_play = play_adapted.join(game_adapted, on=["gameId", "TeamAbbr"], how = "left")

import numpy as np

def get_defense(arr, ref):
  if arr[4] == "home":
    filtered = (ref.
                filter(pl.col("gameId") == arr[0]).
                filter(pl.col("team") == "visitor"))
    return filtered.to_numpy()
  else:
    filtered = (ref.
                filter(pl.col("gameId") == arr[0]).
                filter(pl.col("team") == "home"))
    return filtered.to_numpy()

def get_second_team(arr, ref):
  second = get_defense(arr, ref)[0]
  to_stack = np.array([arr[0], arr[1], second[2], "Defense", second[1]])
  return np.stack([arr, to_stack])

def get_new_arr(df, ref):
  arr = np.stack([get_second_team(v, ref) for v in df.to_numpy()])
  arr = np.reshape(arr, (arr.shape[0]*arr.shape[1], arr.shape[2]))
  return arr

new_arr = get_new_arr(game_play, game_adapted)
new_dict = {game_play.columns[i] : [r[i] for r in new_arr] for i in range(len(game_play.columns))}

new_dict["gameId"] = [int(v) for v in new_dict["gameId"]]
new_dict["playId"] = [int(v) for v in new_dict["playId"]]
new_dict["TeamAbbr"] = [str(v) for v in new_dict["TeamAbbr"]]
new_dict["OffDef"] = [str(v) for v in new_dict["OffDef"]]
new_dict["team"] = [str(v) for v in new_dict["team"]]

game_play = pl.from_dict(new_dict)

cols = ["gameId", "playId", "team", "nflId", "position", "frameId", "x", "y"]

new_data_final = (data.
                  filter(pl.col("team") != "football").
                  with_columns(pl.col("team").str.replace("away", "visitor")).
                  select(cols).
                  join(game_play, on = ["gameId", "playId", "team"], how = "left").
                  drop("team").
                  rename({"TeamAbbr": "team"}).
                  with_columns(pl.lit("Passing").alias("PlayType")).
                  select(["PlayType", "gameId", "playId", "team", "OffDef", "nflId", "position", "frameId", "x", "y"]).
                  with_columns((pl.col("frameId")%2).alias("filter")).
                  filter(pl.col("filter") == 1).
                  drop("filter"))

new_data_final.shape

new_data_final.write_parquet(path+"/processed_df.parquet")

data = pl.read_parquet(path+"/processed_df.parquet")

years_to_get = [2017, 2018, 2019, 2020, 2021, 2022]
rosters = pl.from_pandas(nfl.import_rosters(years_to_get))
season_data = pl.from_pandas(nfl.import_pbp_data(years_to_get))
players = pl.read_csv(path+"/players.csv")

data.columns

filter((pl.col("gameId") != 2018093011) &
                       (~pl.col("nflId").is_in(["43336", "38852"]))

updated_data = (data.
                groupby("PlayType", "gameId", "playId", "team", "OffDef", "nflId", "position").
                agg([
                    pl.col("frameId"),
                    pl.col("x"),
                    pl.col("y")
                ]).
                with_columns(pl.col("nflId").cast(pl.Int64)).
                join(
                    players.
                    select("nflId", "displayName").
                    rename({"displayName" : "player_name"}),
                    on = "nflId",
                    how = "left").
                join(
                    rosters.
                    filter(pl.col("season") == 2018).
                    select("team", "player_name", "gsis_it_id").
                    rename({"team" : "check_team"}),
                    on = "player_name",
                    how = "left").
                drop("nflId").
                rename({"gsis_it_id" : "nflId"}).
                filter((pl.col("gameId") != 2018093011) &
                       (~pl.col("nflId").is_in(["43336", "38852"])))
                )

checks = (updated_data.
          select("gameId", "playId", "nflId").
          groupby("gameId", "playId", "nflId").
          count().
          rename({"count" : "NbOcc"}))

checks.select("NbOcc").unique()

nb_players_per_play = (updated_data.
                       select("gameId", "playId", "nflId").
                       groupby("gameId", "playId").
                       count().
                       rename({"count" : "nb_players_check"}))

final_data = (updated_data.
              join(nb_players_per_play,
                   on = ["gameId", "playId"],
                   how = "left").
              filter(pl.col("nb_players_check") < 23).
              drop("nb_players_check").
              explode(["frameId", "x", "y"]))

final_data = (final_data.
              select("PlayType", "gameId", "playId", "team", "OffDef", "nflId", "position", "frameId", "x", "y"))

final_data.write_parquet(path+"/processed_df.parquet")

final_data.filter(pl.col("nflId").is_null())

updated_data.filter(pl.col("nflId").is_null())
