==> notebooks/QBGPT/0_dataset_creation_analytics.ipynb <==
!pip install --upgrade polars
!pip install nfl_data_py

import os
import pandas as pd
import polars as pl
import nfl_data_py as nfl

from google.colab import drive
drive.mount('/content/gdrive')

os.chdir("/content/gdrive/MyDrive/NFL_Challenge/NFL-GPT/NFL data")

os.listdir()

**2023 Data Bowl**

path = "asonty"
os.listdir(path)[0:10]

import csv
import numpy as np

def get_data(path):
  # open .tsv file
  with open(path) as file:
      tsv_file = csv.reader(file, delimiter="\t")
      lines = [line for line in tsv_file]
      df = pd.DataFrame(np.array(lines[1:]), columns = lines[0])
      return pl.from_pandas(df)

to_keep = ['gameId', 'playId', 'playType', 'teamAbbr', 'possessionFlag', 'gsisId', 'position', 'frame', 'x', 'y']

def process_data(df, selected_cols = to_keep):
  selected =  (df.
              filter(pl.col("displayName") != "ball").
              select(selected_cols).
               with_columns(
                  [pl.col("gameId").cast(pl.Float64).cast(pl.Int64),
                   pl.col("playId").cast(pl.Float64).cast(pl.Int64),
                   pl.col("possessionFlag").cast(pl.Float64),
                   pl.col("frame").cast(pl.Float64).cast(pl.Int64),
                   pl.col("x").cast(pl.Float64),
                   pl.col("y").cast(pl.Float64)]
               ).
              rename({"playType": "PlayType",
                      "teamAbbr": "team",
                      "frame": "frameId"}).
             with_columns(
                pl.when(pl.col("possessionFlag") == 1).
                then(pl.lit("Offense")).
                otherwise(pl.lit("Defense")).
                alias("OffDef")).
             drop("possessionFlag").
             with_columns(
                (pl.col("frameId") % 2).alias("filter")
             ).
             filter(pl.col("filter") == 1).
             drop("filter").
             with_columns(
                pl.col("PlayType").str.replace("play_type_pass", "Passing")).
             with_columns(
                pl.col("PlayType").str.replace("play_type_rush", "Rushing")).
             with_columns(
                pl.col("PlayType").str.replace("play_type_sack", "Passing")).
             with_columns(
                pl.col("PlayType").str.replace("play_type_kickoff", "Kickoff")).
             with_columns(
                pl.col("PlayType").str.replace("play_type_punt", "Punt")).
             with_columns(
                pl.col("PlayType").str.replace("play_type_two_point_conversion", "two_point")).
            select("PlayType", "gameId", "playId", "team", "OffDef", "gsisId", "position", "frameId", "x", "y"))

  return selected

def data_extract(path, cols = to_keep):
  opened = get_data(path)
  processed = process_data(opened, cols)
  return processed

index_path = ["ngs_highlights_index.tsv", "processed_df.parquet"]
to_read = [p for p in os.listdir(path) if p not in index_path]

from tqdm import tqdm
progress_bar = tqdm(to_read, desc="reading", ncols=100)

new_data_final = pl.concat([data_extract("asonty/"+path) for path in progress_bar])

progress_bar.close()

new_data_final.shape

years_to_get = [2017, 2018, 2019, 2020, 2021, 2022]
rosters = pl.from_pandas(nfl.import_rosters(years_to_get))

new_data_final = (new_data_final.
                  with_columns(pl.col("gameId").cast(pl.Utf8).str.slice(0, 4).cast(pl.Int32).alias("season")).
                  join((rosters.
                        select("season", "player_id", "gsis_it_id").
                        rename({"player_id" : "gsisId", "gsis_it_id": "nflId"})),
                       on = ["season", "gsisId"],
                       how = "left").
                  drop("gsisId").
                  select("PlayType", "gameId", "playId", "team", "OffDef", "nflId", "position", "frameId", "x", "y"))

new_data_final.write_parquet(path+"/processed_df.parquet")
