==> notebooks/QBGPT/0_dataset_creation_NGS_H2019.ipynb <==
!pip install polars nfl_data_py

import os
import pandas as pd
import polars as pl
import nfl_data_py as nfl

from google.colab import drive
drive.mount('/content/gdrive')

os.chdir("/content/gdrive/MyDrive/NFL_Challenge/NFL-GPT/NFL data")

os.listdir()

season_2019 = nfl.import_pbp_data([2019, 2020])

def format_ref(df):
  selected = df[["old_game_id", "play_id", "play_type", "home_team", "away_team", "posteam", "posteam_type"]].melt(id_vars = ["old_game_id",	"play_id",	"play_type", "posteam",	"posteam_type"], var_name = "team", value_name = "TeamAbbr")
  selected["team"] = [v.replace("_team", "") for v in selected["team"]]
  selected = selected[(selected["play_type"] != "no_play") & (selected["play_type"].notnull())].reset_index(drop = True)
  selected["old_game_id"] = selected["old_game_id"].astype(int)
  selected["play_id"] = selected["play_id"].astype(int)

  selected["to_merge"] = selected["posteam"] == selected["TeamAbbr"]
  selected = selected.drop("posteam", axis = "columns").rename({"to_merge": "possTeam"}, axis = "columns")
  selected["OffDef"] = ["Offense" if v is True else "Defense" for v in selected["possTeam"]]
  selected = selected.drop(["posteam_type",	"team"], axis = "columns").rename({"TeamAbbr": "team", "old_game_id" : "game_id"}, axis = "columns")
  return selected

ref_season = format_ref(season_2019)

ref_season

ref_season[ref_season["game_id"] == 2020010501]

**2019 concat**

from datetime import datetime

def convert_time(time):
    return datetime.strptime(time, "%Y-%m-%d %H:%M:%S.%f")

path = "Highlights_NGS_2019"

from datetime import datetime

def insert_frame_id(df):
  df.insert(1, "frameId", range(1, df.shape[0]+1))
  return df

def get_data_from_csv(path, reference):
  print(path)
  data = pd.read_csv(path, index_col = "Unnamed: 0")
  filtered = data[data["displayName"] != "ball"].reset_index(drop = True)
  filtered["time"] = filtered["time"].apply(convert_time)

  possible_frames = filtered[["play_id", "time"]].drop_duplicates().sort_values("time").reset_index(drop = True)
  frame_dfs = [possible_frames[possible_frames["play_id"] == v] for v in possible_frames["play_id"].unique()]
  frame_dfs = [insert_frame_id(d) for d in frame_dfs]
  final_df = pd.concat(frame_dfs).reset_index(drop = True)

  final = pd.merge(filtered, final_df, on = ["play_id", "time"], how = "left")
  final["game_id"] = final["game_id"].astype(int)

  if path == "Highlights_NGS_2019/Highlight_19_post.csv":
    final["possTeam"] = final["IsOnOffense"]

  final = final.merge(reference, on = ["game_id", "play_id", "possTeam"], how = "left")[["play_type", "game_id", "play_id", "team", "OffDef", "gsisId", "position", "frameId", "x", "y"]].rename({"play_type": "PlayType", "game_id": "gameId", "play_id": "playId"}, axis = "columns")

  if path == "Highlights_NGS_2019/Highlight_19_post.csv":
    final = final.dropna().reset_index(drop = True)

  final["PlayType"] = final["PlayType"].replace({"kickoff": "Kickoff",
                                                "pass": "Passing",
                                                "run": "Rushing",
                                                "punt": "Punt",
                                                "extra_point": "Extra Point",
                                                "qb_kneel": "Kneel",
                                                "field_goal": "Field Goal",
                                                "qb_spike": "Spike"})



  print(path)
  print(sum(final.isna().sum()))

  return pl.from_pandas(final)

to_read_2019 = os.listdir(path)
avoid = ["Highlight_19_week1.csv", "Highlight_19_week21.csv", "processed_df.parquet"]

to_read_2019 = [v for v in to_read_2019 if v not in avoid]

new_data_final = pl.concat([get_data_from_csv("Highlights_NGS_2019/"+path, ref_season) for path in to_read_2019])
new_data_final = (new_data_final.
                  with_columns(
                      (pl.col("frameId")%2).alias("filter")).
                  filter(pl.col("filter")==1).
                  drop("filter"))

new_data_final.shape

years_to_get = [2017, 2018, 2019, 2020, 2021, 2022]
rosters = pl.from_pandas(nfl.import_rosters(years_to_get))

new_data_final = (new_data_final.
                  with_columns(pl.col("gameId").cast(pl.Utf8).str.slice(0, 4).cast(pl.Int32).alias("season")).
                  join((rosters.
                        select("season", "player_id", "gsis_it_id").
                        rename({"player_id" : "gsisId", "gsis_it_id": "nflId"})),
                       on = ["season", "gsisId"],
                       how = "left").
                  drop("gsisId").
                  select("PlayType", "gameId", "playId", "team", "OffDef", "nflId", "position", "frameId", "x", "y"))

new_data_final.write_parquet(path+"/processed_df.parquet")

path
