==> notebooks/QBGPT/1_data_concat.ipynb <==
!pip install --upgrade polars
!pip install nfl_data_py

import os
import pandas as pd
import polars as pl
import nfl_data_py as nfl


from google.colab import drive
drive.mount('/content/gdrive')

os.chdir("/content/gdrive/MyDrive/NFL_Challenge/QB-GPT/data_preprocessing/0_raw")

os.listdir()

to_get = [
 'data bowl 2021',
 'data bowl 2022',
 'data bowl 2023',
 'data bowl 2024',
 'asonty',
 'Highlights_NGS_2019']

def read(p):
  data = pl.read_parquet(p)
  to_insert = p.replace("/processed_df.parquet", "")
  data = (data.
          with_columns(pl.col("nflId").cast(pl.Int64)).
          with_columns(pl.lit(to_insert).alias("data_source")))
  return data

years_to_get = [2018, 2019, 2020, 2021, 2022, 2023]
season_data = pl.from_pandas(nfl.import_pbp_data(years_to_get))
rosters = pl.from_pandas(nfl.import_seasonal_rosters(years_to_get))

dataset = pl.concat([read(p+"/processed_df.parquet").unique() for p in to_get])

(dataset.
 head())

dataset = (dataset.
           with_columns(pl.when(pl.col("team") == "OAK").
                        then(pl.lit("LV")).
                        otherwise(pl.col("team"))).
           group_by('data_source', 'PlayType', 'gameId', 'playId', 'team', 'OffDef', 'nflId', 'position').
           agg([pl.col("frameId"), pl.col("x"), pl.col("y")]).
           filter(pl.col("nflId").is_not_null()).
           join(
              season_data.
              select("old_game_id", "play_id", "posteam").
              unique().
              with_columns([
                  pl.col("old_game_id").cast(pl.Int64).alias("gameId"),
                  pl.col("play_id").cast(pl.Int64).alias("playId")
              ]).
              drop("old_game_id", "play_id"),
              on = ["gameId", "playId"],
              how = "left"
           ).
           with_columns(pl.when(pl.col("posteam") == pl.col("team")).
                        then(pl.lit("Offense")).
                        otherwise(pl.lit("Defense")).
                        alias("new_team")).
           drop("OffDef", "posteam").
           rename({"new_team" : "OffDef"}).
           join(season_data.
                select("old_game_id", "play_id", "yardline_100", "qtr", "down", "play_type", "season_type", "season").
                with_columns([
                    pl.col("old_game_id").cast(pl.Int64).alias("gameId"),
                    pl.col("play_id").cast(pl.Int64).alias("playId")
                    ]).
                drop("old_game_id", "play_id"),
                on = ["gameId", "playId"],
                how = "left").
           drop("PlayType").
           rename({"play_type" : "PlayType"}).
           with_columns(pl.col("frameId").list.lengths().alias("NB_Frames")).
           select("data_source", "season_type", "season", "gameId", "qtr", "down", "yardline_100", "playId", "PlayType", "team", "OffDef", "nflId", "position", "frameId", "x", "y").
           unique(subset = ["gameId", "playId", "nflId"]))


filter_NB_players = (dataset.
                     select("gameId", "playId", "OffDef", "nflId").
                     unique().
                     group_by("gameId", "playId", "OffDef").
                     count().
                     rename({"count" : "Nb_players"}))

dataset = (dataset.
           join(filter_NB_players,
                on = ["gameId", "playId", "OffDef"],
                how = "left").
           filter(pl.col("Nb_players") > 1).
           drop("Nb_players"))

import gc
_ = gc.collect()

uniques = (dataset.
           select("gameId", "playId", "nflId").
           unique())

dataset = (uniques.
           join(dataset,
                on = ["gameId", "playId", "nflId"]))

## V1

specific_to_add = (dataset.
                   select("data_source", "season", "gameId", "playId", "OffDef", "nflId").
                   group_by("data_source", "season", "gameId", "playId", "OffDef").
                   agg(pl.col("nflId").cast(pl.Utf8).alias("players_scraped")).
                   join(
                      season_data.
                      with_columns(
                          [
                              pl.col("old_game_id").cast(pl.Int64).alias("gameId"),
                              pl.col("play_id").cast(pl.Int64).alias("playId")
                              ]).
                      select("season", "gameId", "playId", "offense_players", "defense_players").
                      melt(id_vars = ["season", "gameId", "playId"],
                           value_vars = ["offense_players", "defense_players"],
                           variable_name = "OffDef",
                           value_name = "player_id").
                      with_columns(pl.when(pl.col("OffDef") == "offense_players").
                                   then(pl.lit("Offense")).
                                   otherwise(pl.lit("Defense")).
                                   alias("OffDef")).
                      with_columns(pl.col("player_id").str.split(";")).
                      explode("player_id").
                      join(
                          rosters.
                          select("player_id", "gsis_it_id").
                          unique().
                          rename({"gsis_it_id" : "nflId"}),
                          on = "player_id",
                          how = "left"
                      ).
                      drop("player_id").
                      group_by("season", "gameId", "playId", "OffDef").
                      agg(pl.col("nflId").alias("players_present")),
                      on = ["season", "gameId", "playId", "OffDef"],
                      how = "left").
                   with_columns(pl.col("players_present").list.set_difference("players_scraped").alias("difference")).
                   select("data_source", "season", "gameId", "playId", "OffDef", "difference").
                   explode("difference").
                   filter(pl.col("difference") != "").
                   rename({"difference" : "nflId"}).
                   with_columns(pl.col("nflId").cast(pl.Int64)).
                   join(
                      rosters.
                      select("season", "gsis_it_id", "position").
                      with_columns([
                          pl.col("season").cast(pl.Int64).alias("season"),
                          pl.col("gsis_it_id").cast(pl.Int64).alias("nflId")
                          ]).
                      drop("gsis_it_id").
                      unique(subset=["season", "nflId"]),
                      on = ["season", "nflId"],
                      how = "left").
                   fill_null("UNKNOWN").
                   join(
                      dataset.
                      select("season", "gameId", "playId", "season_type", "team", "OffDef", "qtr", "down", "yardline_100", "PlayType").
                      unique(),
                      on = ["season", "gameId", "playId", "OffDef"],
                      how = "left").
                   select('data_source', 'season_type', 'season', 'gameId', 'qtr', 'down', 'yardline_100', 'playId', 'PlayType', 'team', 'OffDef', 'nflId', 'position').
                   with_columns([
                      pl.lit(-1).cast(pl.Int64).alias("frameId"),
                      pl.lit(-1).cast(pl.Float64).alias("x"),
                      pl.lit(-1).cast(pl.Float64).alias("y")
                   ]).
                   group_by('data_source', 'season_type', 'season', 'gameId', 'qtr', 'down', 'yardline_100', 'playId', 'PlayType', 'team', 'OffDef', 'nflId', 'position').
                   agg([pl.col("frameId"),
                        pl.col("x"),
                        pl.col("y")]).
                   with_columns(pl.lit(-1).alias("Length")))

specific_to_keep = (dataset.
                    select("data_source", "season", "gameId", "playId", "nflId").
                    group_by("data_source", "season", "gameId", "playId").
                    agg(pl.col("nflId").cast(pl.Utf8).alias("players_scraped")).
                    join(
                        season_data.
                        with_columns(
                            [
                                pl.col("old_game_id").cast(pl.Int64).alias("gameId"),
                                pl.col("play_id").cast(pl.Int64).alias("playId")
                                ]).
                        select("gameId", "playId", "players_on_play").
                        with_columns(pl.col("players_on_play").str.split(";").alias("players_present")),
                        on = ["gameId", "playId"],
                        how = "left").
                    drop("players_on_play").
                    with_columns(pl.col("players_present").list.set_intersection("players_scraped").alias("kept")).
                    select("data_source", "season", "gameId", "playId", "kept").
                    explode("kept").
                    filter(pl.col("kept") != "").
                    rename({"kept" : "nflId"}).
                    with_columns(pl.col("nflId").cast(pl.Int64)).
                    join(
                        dataset,
                        on = ["season", "gameId", "playId", "nflId"],
                        how = "left").
                    select('data_source', 'season_type', 'season', 'gameId', 'qtr', 'down', 'yardline_100', 'playId', 'PlayType', 'team', 'OffDef', 'nflId', 'position', 'frameId', 'x', 'y').
                    with_columns(pl.col("frameId").list.lengths().cast(pl.Int32).alias("Length")))

dataset = (pl.concat([specific_to_keep, specific_to_add]).
           filter(pl.col("Length") != -1))

(dataset.
 select("gameId", "playId", "team", "OffDef", "nflId").
 unique().
 shape)

(dataset.head())

dataset.write_parquet("final_df_2024.parquet")
