==> notebooks/QBGPT/4_training.ipynb <==
!pip install tensorflow==2.14

import os
import pandas as pd
import numpy as np
import tensorflow as tf

env = "drive"

if env != "local":
  !git clone https://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@github.com/samchaineau/QB-GPT.git
  import sys
  sys.path.append("/content/QB-GPT/")

if env == "local":
    os.chdir("/Users/samuel/Documents/GitHub/QB-GPT/")
else:
    from google.colab import drive
    drive.mount('/content/gdrive')
    os.chdir("/content/gdrive/MyDrive/NFL_Challenge/QB-GPT/")

os.listdir()

training_data = tf.data.Dataset.load("data_models/QBGPT/train_tokens_NFL_GPT_v2")
testing_data = tf.data.Dataset.load("data_models/QBGPT/test_tokens_NFL_GPT_v2")

train_length = [i for i,_ in enumerate(training_data)][-1] + 1
test_length = [i for i,_ in enumerate(testing_data)][-1] + 1

print("Train length is : ", str(train_length))
print("Test length is : ", str(test_length))

batch_size = 32

training_data = training_data.shuffle(train_length).batch(batch_size)
testing_data = testing_data.shuffle(test_length).batch(batch_size)

## Model classes

os.listdir("models/modeling/QBGPT/")

def insert_weights(df, w):
  df["weights"] = [w for i in range(df.shape[0])]
  return df

class_weights = pd.read_parquet("models/modeling/QBGPT/class_weightsv2.parquet")

step_range = [(0, 10), (10, 100), (100, 1000), (1000, 10000), (10000, 50000), (50000, 100000), (100000, 300000), (300000, 500000), (500000, 1000000), (1000000, 10000000)]

from collections import Counter
weights = dict(Counter(class_weights["Zone_ID"].to_numpy()))
weights_df = pd.DataFrame(np.array([[k, v] for k,v in weights.items()]), columns = ["Class", "Count"])

weights_dict = {i : weights_df[(weights_df['Count'] > step_range[i][0]) & (weights_df['Count'] <= step_range[i][1])].reset_index(drop = True) for i in range(len(step_range))}
w_dict = {0 : 1,
          1 : 0.9,
          2 : 0.8,
          3 : 0.7,
          4 : 0.6,
          5 : 0.5,
          6 : 0.4,
          7 : 0.3,
          8 : 0.2,
          9 : 0.05,}

weights_dict = {k:insert_weights(v, w_dict[k]) for k,v in weights_dict.items()}

weights_df = pd.concat(list(weights_dict.values())).reset_index(drop = True)

weights_inv = {v[0] : v[2] for v in weights_df.values}

max(weights_inv.keys())

def scheduler(epoch, lr):
  if epoch < 1:
    return 3e-3
  elif (epoch >= 1) & (epoch < 2):
    return 2e-3
  elif (epoch >= 2) & (epoch < 3):
    return 1e-3
  elif (epoch >= 3) & (epoch < 5):
    return 5e-4
  elif (epoch >= 5) & (epoch < 7):
    return 1e-4
  else:
    return 5e-5


schedule = tf.keras.callbacks.LearningRateScheduler(scheduler)

from models.modeling.QBGPT.models import QBGPT, LargeQBGPT, XLargeQBGPT
from models.modeling.QBGPT.losses_and_metrics import CustomSparseCategoricalAccuracy, CustomTopKAccuracy, CustomSparseCategoricalCrossentropy

moves_to_pred = 11221
input_size = 11223
starts_size = 2114
scrimmage_size = 100
positions_id = 29

temp_ids = 52
off_def_size = 2
token_type_size = 3
play_type_size = 9

model_large = LargeQBGPT(input_vocab_size = input_size,
                         positional_vocab_size = temp_ids,
                         position_vocab_size=positions_id,
                         start_vocab_size=starts_size,
                         scrimmage_vocab_size=scrimmage_size,
                         offdef_vocab_size = off_def_size,
                         type_vocab_size = token_type_size,
                         playtype_vocab_size = play_type_size,
                         embedding_dim = 128,
                         hidden_dim = 128,
                         num_heads = 3,
                         diag_masks = True,
                         to_pred_size = moves_to_pred)

model_medium = QBGPT(input_vocab_size = input_size,
                    positional_vocab_size = temp_ids,
                    position_vocab_size=positions_id,
                    start_vocab_size=starts_size,
                    scrimmage_vocab_size=scrimmage_size,
                    offdef_vocab_size = off_def_size,
                    type_vocab_size = token_type_size,
                    playtype_vocab_size = play_type_size,
                    embedding_dim = 256,
                    hidden_dim = 256,
                    num_heads = 3,
                    diag_masks = True,
                    to_pred_size = moves_to_pred)

model_small = QBGPT(input_vocab_size = input_size,
                    positional_vocab_size = temp_ids,
                    position_vocab_size=positions_id,
                    start_vocab_size=starts_size,
                    scrimmage_vocab_size=scrimmage_size,
                    offdef_vocab_size = off_def_size,
                    type_vocab_size = token_type_size,
                    playtype_vocab_size = play_type_size,
                    embedding_dim = 128,
                    hidden_dim = 128,
                    num_heads = 3,
                    diag_masks = True,
                    to_pred_size = moves_to_pred)

model_tiny = QBGPT(input_vocab_size = input_size,
                    positional_vocab_size = temp_ids,
                    position_vocab_size=positions_id,
                    start_vocab_size=starts_size,
                    scrimmage_vocab_size=scrimmage_size,
                    offdef_vocab_size = off_def_size,
                    type_vocab_size = token_type_size,
                    playtype_vocab_size = play_type_size,
                    embedding_dim = 64,
                    hidden_dim = 64,
                    num_heads = 3,
                    diag_masks = True,
                    to_pred_size = moves_to_pred)

custom_loss = CustomSparseCategoricalCrossentropy(from_logits=True, class_weights=weights_inv)

model_tiny.compile(optimizer=tf.keras.optimizers.Adam(),
                    loss=custom_loss,
                    metrics=[CustomSparseCategoricalAccuracy(),
                             CustomTopKAccuracy(k=3, name='custom_top_3_accuracy'),
                             CustomTopKAccuracy(k=5, name='custom_top_5_accuracy')])

history_tiny = model_tiny.fit(training_data, validation_data = testing_data, epochs=9, callbacks = [schedule])

pd.DataFrame(history_tiny.history).to_csv("models/modeling/QBGPT/history/training_history_model_tinyv2.csv", index = False, sep = ";")

model_tiny.save_weights("models/modeling/QBGPT/weights/model_tinyv2/QBGPT")

model_tiny.summary()

custom_loss = CustomSparseCategoricalCrossentropy(from_logits=True, class_weights=weights_inv)

model_small.compile(optimizer=tf.keras.optimizers.Adam(),
                    loss=custom_loss,
                    metrics=[CustomSparseCategoricalAccuracy(),
                             CustomTopKAccuracy(k=3, name='custom_top_3_accuracy'),
                             CustomTopKAccuracy(k=5, name='custom_top_5_accuracy')])

history_small = model_small.fit(training_data, validation_data = testing_data, epochs=9, callbacks = [schedule])

pd.DataFrame(history_small.history).to_csv("models/modeling/QBGPT/history/training_history_model_smallv2.csv", index = False, sep = ";")

model_small.save_weights("models/modeling/QBGPT/weights/model_smallv2/QBGPT")

model_small.summary()

custom_loss = CustomSparseCategoricalCrossentropy(from_logits=True, class_weights=weights_inv)

model_medium.compile(optimizer=tf.keras.optimizers.Adam(),
                     loss=custom_loss,
                     metrics=[CustomSparseCategoricalAccuracy(),
                              CustomTopKAccuracy(k=3, name='custom_top_3_accuracy'),
                              CustomTopKAccuracy(k=5, name='custom_top_5_accuracy')])

history_medium = model_medium.fit(training_data, validation_data = testing_data, epochs=9, callbacks = [schedule])

pd.DataFrame(history_medium.history).to_csv("models/modeling/QBGPT/history/training_history_model_mediumv2.csv", index = False, sep = ";")

model_medium.save_weights("models/modeling/QBGPT/weights/model_mediumv2/QBGPT")

model_medium.summary()

def scheduler(epoch, lr):
  if epoch < 1:
    return 5e-3
  elif (epoch >= 1) & (epoch < 2):
    return 2e-3
  elif (epoch >= 2) & (epoch < 3):
    return 1e-3
  elif (epoch >= 3) & (epoch < 5):
    return 5e-4
  elif (epoch >= 5) & (epoch < 7):
    return 1e-4
  else:
    return 5e-5


schedule = tf.keras.callbacks.LearningRateScheduler(scheduler)

custom_loss = CustomSparseCategoricalCrossentropy(from_logits=True, class_weights=weights_inv)

model_large.compile(optimizer=tf.keras.optimizers.Adam(),
                    loss=custom_loss,
                    metrics=[CustomSparseCategoricalAccuracy(),
                             CustomTopKAccuracy(k=3, name='custom_top_3_accuracy'),
                             CustomTopKAccuracy(k=5, name='custom_top_5_accuracy')])

history_large = model_large.fit(training_data, validation_data = testing_data, epochs=9, callbacks = [schedule])

pd.DataFrame(history_large.history).to_csv("models/modeling/QBGPT/history/training_history_model_largev2.csv", index = False, sep = ";")

model_large.save_weights("models/modeling/QBGPT/weights/model_largev2/QBGPT")

model_large.summary()
